AGHAMazingQuestMobile/requirements.txt
Django==5.0
djangorestframework==3.15
djangorestframework-simplejwt==5.3
dj-rest-auth==6.0
django-allauth==0.63
drf-spectacular==0.27
psycopg[binary]==3.2.1
django-cors-headers==4.3
celery==5.3
redis==5.0
gunicorn==23.0.0
uvicorn==0.22.0
dj-database-url==1.0.0
python-dotenv==1.0.0 
requests
cryptography
Pillow==10.1.0